requirements.txt
# Core Framework
fastapi==0.135.1
starlette==0.49.1
uvicorn[standard]==0.27.0
pydantic==2.11.7
python-multipart==0.0.22

# Elasticsearch
elasticsearch==8.11.0

# Natural Language Processing
spacy==3.8.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# AI/LLM
openai==2.26.0
tiktoken==0.12.0

# Database
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.17.0

# Caching
redis==6.4.0

# Async HTTP
aiohttp==3.13.3
urllib3==2.6.3

# Environment
python-dotenv==1.1.1

# CORS
fastapi-cors==0.0.6
